Reorganizing the collections and documents.
