+++ grow/common/json_encoder.py
@@ -2,8 +2,8 @@
 
 import datetime
 import json
-from grow.pods import document_fields
-from grow.pods import documents
+from grow.documents import documents
+from grow.documents import document_fields
 
 
 class GrowJSONEncoder(json.JSONEncoder):
