+++ grow/documents/document_format_test.py
@@ -1,10 +1,11 @@
-import unittest
+"""Tests for document fields."""
 
-from . import document_format
-from . import pods
-from . import storage
-from grow.testing import testing
 import textwrap
+import unittest
+from grow.documents import document_format
+from grow.pods import pods
+from grow.pods import storage
+from grow.testing import testing
 
 
 class DocumentFormatTestCase(unittest.TestCase):
