+++ grow/templates/tags.py
@@ -5,8 +5,8 @@ import itertools
 from datetime import datetime
 import jinja2
 from babel.messages import catalog as babel_catalog
+from grow.collections import collection as collection_lib
 from grow.common import utils
-from grow.pods import collection as collection_lib
 from grow.pods import locales as locales_lib
